Extract portfolio performance snapshot contracts
