+++ src/app/routers/portfolio_performance.py
@@ -3,7 +3,9 @@ from typing import Annotated
 
 from fastapi import APIRouter, Depends, Query
 
-from app.contracts.portfolio import PortfolioPerformanceSnapshotResponse
+from app.contracts.portfolio_performance_snapshot import (
+    PortfolioPerformanceSnapshotResponse,
+)
 from app.middleware.correlation import correlation_id_var
 from app.services.portfolio_service_provider import portfolio_performance_workspace_service
